ADM-359 rename and move useAppDispatch file seat
