+++ frontend/src/components/Metrics/MetricsStepper/index.tsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box'
 import Stepper from '@mui/material/Stepper'
 import Step from '@mui/material/Step'
 import { MetricsStepperContent, MetricsStepLabel } from './style'
-import { useAppSelector } from '@src/hooks'
+import { useAppSelector } from '@src/hooks/useAppDispatch'
 import { selectStep } from '@src/features/stepper/StepperSlice'
 import { ConfigStep } from '@src/components/Metrics/ConfigStep'
 import { STEPS } from '@src/constants'
